Agent/requirements.txt
openai
langchain
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
load_dotenv
streamlit
langchain-openai
langchain-community